requirements.txt
alpaca-py==0.7.1
certifi==2022.12.7
charset-normalizer==3.0.1
chess==1.9.4
click==8.1.3
Flask==2.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.4
numpy==1.23.3
pandas==1.5.3
pydantic==1.10.4
pygame==2.1.2
python-chess==1.999
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.9.1
six==1.16.0
sseclient-py==1.7.2
termcolor==1.1.0
typing_extensions==4.4.0
urllib3==1.26.14
websockets==10.4
Werkzeug==2.2.2
